Rename Alert comp to Dialog
