+++ src/prompts/dialog/banner.scss
