+++ src/prompts/prompts.scss
@@ -6,5 +6,5 @@
 @import './styles/buttons';
 
 @import './bell/bell.scss';
-@import './alert/alert.scss';
-@import './alert/banner.scss';
+@import './dialog/alert.scss';
+@import './dialog/banner.scss';
